fixed bugs

+++ backend/api/views.py
@@ -19,6 +19,8 @@ import psycopg2
 import mysql.connector
 from datetime import timedelta
 from django.utils import timezone 
+from sqlalchemy import create_engine
+
 
 
 
@@ -204,31 +206,22 @@ class UserDatabaseView(generics.GenericAPIView):
                 temp_file_path = 'temp.sqlite'
 
                 # Write uploaded file to a temporary file
-                with open('temp.sqlite', 'wb+') as temp_file:
+                with open(temp_file_path, 'wb+') as temp_file:
                     for chunk in file.chunks():
                         temp_file.write(chunk)
                 # Connect to the temporary SQLite database file
-                conn = sqlite3.connect('temp.sqlite')
+                conn = sqlite3.connect(temp_file_path)
+                df = pd.read_sql_query(f"SELECT * FROM {table}", conn)
+                conn.close()
             elif db_type == 'mysql':
-                conn = mysql.connector.connect(
-                    host=host,
-                    port=port,
-                    database=dbname,
-                    user=user,
-                    password=password
-                )
+                engine = create_engine(f'mysql+mysqlconnector://{user}:{password}@{host}:{port}/{dbname}')
+                df = pd.read_sql_query(f"SELECT * FROM {table}", engine)
             elif db_type == 'postgresql':
-                conn = psycopg2.connect(
-                    dbname=dbname,
-                    user=user,
-                    password=password,
-                    host=host,
-                    port=port
-                )
+                engine = create_engine(f'postgresql+psycopg2://{user}:{password}@{host}:{port}/{dbname}')
+                df = pd.read_sql_query(f"SELECT * FROM {table}", engine)
             else:
                 return Response({'error': 'Unsupported database type'}, status=status.HTTP_400_BAD_REQUEST)
 
-            df = pd.read_sql_query(f"SELECT * FROM {table}", conn)
             # Convert columns to numeric if possible
             numeric_columns = df.select_dtypes(include='object').columns
             for col in numeric_columns:
@@ -240,11 +233,10 @@ class UserDatabaseView(generics.GenericAPIView):
             # Generate charts
             charts_base64 = generate_charts(df)
             
-            # Close the database connection
-            conn.close()
             # Delete temporary file
-            if os.path.exists(temp_file_path):
+            if db_type == 'sqlite' and os.path.exists(temp_file_path):
                 os.remove(temp_file_path)
+
             return Response({'charts': charts_base64}, status=status.HTTP_200_OK)
         except Exception as e:
             return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
